Add 🚒: Avances con administracion de IO

+++ utils/headers/utils.h
@@ -17,7 +17,7 @@
 
 bool config_has_all_properties(t_config*, char**);
 
-bool send_un_char_y_un_int(int fd, char* cadena, uint8_t entero);
+bool send_un_char_y_un_int(int fd, char* cadena, uint16_t entero);
 bool send_dos_ints(int fd, uint8_t int1, uint8_t int2);
 
 bool recv_un_char_y_un_int(int fd, char** cadena, uint8_t* entero);
